requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.11.2
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.38.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.31.3
iopath==0.1.10
jax==0.4.8
joblib==1.2.0
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
ml-dtypes==0.1.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.0
pandas==2.1.0
pathtools==0.1.2
Pillow==9.4.0
plotly==5.13.0
portalocker==2.7.0
protobuf==4.21.12
psutil==5.9.4
ptflops==0.6.9
pthflops==0.4.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.1
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.8.30
torch==1.13.1
torchmetrics==0.11.1
torchsummary==1.5.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
wandb==0.13.9
Werkzeug==2.3.3
wrapt==1.14.1
yacs==0.1.8
